ddl & counters
