--- internal/models/category.go
@@ -1,34 +1,0 @@
-package models
-
-import (
-	"context"
-	"database/sql"
-	"godiscourse/internal/session"
-)
-
-// topics_count should use pg int64
-const categoriesDDL = `
-CREATE TABLE IF NOT EXISTS categories (
-	category_id           VARCHAR(36) PRIMARY KEY,
-	name                  VARCHAR(36) NOT NULL,
-	alias                 VARCHAR(128) NOT NULL,
-	description           VARCHAR(512) NOT NULL,
-	topics_count          INTEGER NOT NULL DEFAULT 0,
-	last_topic_id         VARCHAR(36),
-	position              INTEGER NOT NULL DEFAULT 0,
-	created_at            TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW(),
-	updated_at            TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW()
-);
-
-CREATE INDEX ON categories (position);
-`
-
-func categoryCount(ctx context.Context, tx *sql.Tx) (int64, error) {
-	var count int64
-	row := tx.QueryRowContext(ctx, "SELECT count(*) FROM categories")
-	err := row.Scan(&count)
-	if err != nil {
-		return 0, session.TransactionError(ctx, err)
-	}
-	return count, nil
-}
