+++ internal/tasks/topic.go
@@ -1,7 +1,6 @@
 package models
 
 import (
-	"context"
 	"database/sql"
 	"fmt"
 	"godiscourse/internal/session"
@@ -9,38 +8,6 @@ import (
 
 	hashids "github.com/speps/go-hashids"
 )
-
-// Topic related CONST
-const (
-	minTitleSize = 3
-	LIMIT        = 50
-)
-
-const topicsDDL = `
-CREATE TABLE IF NOT EXISTS topics (
-	topic_id              VARCHAR(36) PRIMARY KEY,
-	short_id              VARCHAR(255) NOT NULL,
-	title                 VARCHAR(512) NOT NULL,
-	body                  TEXT NOT NULL,
-	comments_count        INTEGER NOT NULL DEFAULT 0,
-	category_id           VARCHAR(36) NOT NULL,
-	user_id               VARCHAR(36) NOT NULL REFERENCES users ON DELETE CASCADE,
-	score                 INTEGER NOT NULL DEFAULT 0,
-	created_at            TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW(),
-	updated_at            TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW()
-);
-CREATE UNIQUE INDEX ON topics (short_id);
-CREATE INDEX ON topics (created_at DESC);
-CREATE INDEX ON topics (user_id, created_at DESC);
-CREATE INDEX ON topics (category_id, created_at DESC);
-CREATE INDEX ON topics (score DESC, created_at DESC);
-`
-
-func topicsCount(ctx context.Context, tx *sql.Tx) (int64, error) {
-	var count int64
-	err := tx.QueryRowContext(ctx, "SELECT count(*) FROM topics").Scan(&count)
-	return count, err
-}
 
 func generateShortID(table string, t time.Time) (string, error) {
 	hd := hashids.NewData()
