+++ internal/user/session.go
@@ -15,16 +15,6 @@ import (
 	"github.com/gofrs/uuid"
 )
 
-const sessionsDDL = `
-CREATE TABLE IF NOT EXISTS sessions (
-	session_id            VARCHAR(36) PRIMARY KEY,
-	user_id               VARCHAR(36) NOT NULL,
-	secret                VARCHAR(1024) NOT NULL,
-	created_at            TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW()
-);
-CREATE INDEX ON sessions (user_id);
-`
-
 // Session contains user's current login infomation
 type Session struct {
 	SessionID string    `sql:"session_id,pk"`
